# backend/src/main/java/de/moehring/coach/buddy/backend/person/dtos/PersonDto.java
package de.moehring.coach.buddy.backend.person.dtos;

import java.time.Instant;
import java.time.LocalDate;
import java.util.List;
import java.util.UUID;

public record PersonDto(
        UUID id,
        String firstName,
        String lastName,
        LocalDate birthDate,
        String nickname,
        String notes,
        List<PhoneNumberDto> phoneNumbers,
        Instant createdAt,
        Instant updatedAt
) {
}