# docs/openapi.json
{
  "openapi" : "3.1.0",
...
  "components" : {
    "schemas" : {
...
      "CreatePersonRelationRequest" : {
        "type" : "object",
        "properties" : {
          "childPersonId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "guardianPersonId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "relationType" : {
            "$ref" : "#/components/schemas/RelationType"
          }
        }
      },
      "CreatePersonRequest" : {
        "type" : "object",
        "required" : [ "firstName", "lastName" ],
        "properties" : {
          "firstName" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "lastName" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "birthDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "nickname" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "phoneNumbers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CreatePhoneNumberRequest"
            }
          }
        }
      },
      "CreatePhoneNumberRequest" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/PhoneType"
          },
          "number" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        },
        "required" : [ "number" ]
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
...
      "PersonDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "birthDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "nickname" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "phoneNumbers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PhoneNumberDto"
            }
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "PersonRelationDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "childPerson" : {
            "$ref" : "#/components/schemas/PersonDto"
          },
          "guardianPerson" : {
            "$ref" : "#/components/schemas/PersonDto"
          },
          "relationType" : {
            "$ref" : "#/components/schemas/RelationType"
          }
        }
      },
      "PhoneNumberDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "type" : {
            "$ref" : "#/components/schemas/PhoneType"
          },
          "number" : {
            "type" : "string"
          }
        }
      },
      "PhoneType" : {
        "type" : "string",
        "enum" : [ "MOBILE", "HOME", "WORK", "EMERGENCY" ]
      },
      "RelationType" : {
        "type" : "string",
        "enum" : [ "MOTHER", "FATHER", "GRANDMOTHER", "GRANDFATHER", "STEP_PARENT", "OTHER" ]
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
      "UpdatePersonRequest" : {
        "type" : "object",
        "required" : [ "firstName", "lastName" ],
        "properties" : {
          "firstName" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "lastName" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "birthDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "nickname" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "phoneNumbers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UpdatePhoneNumberRequest"
            }
          }
        }
      },
      "UpdatePhoneNumberRequest" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/PhoneType"
          },
          "number" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        },
        "required" : [ "number" ]
      }
    }
  },
  "paths" : {
    "/api/v1/person-relations" : {
      "post" : {
        "summary" : "Create person relation",
        "description" : "Creates a relation between a child person and a guardian person.",
        "tags" : [ "Person Relations" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePersonRelationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Person relation created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PersonRelationDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Referenced person not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Person relation already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/person-relations/{id}" : {
      "get" : {
        "summary" : "Find person relation by id",
        "description" : "Returns a single person relation by id.",
        "tags" : [ "Person Relations" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Person relation found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PersonRelationDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Person relation not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete person relation",
        "description" : "Deletes a person relation.",
        "tags" : [ "Person Relations" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Person relation deleted"
          },
          "404" : {
            "description" : "Person relation not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/persons" : {
      "get" : {
        "summary" : "Find persons",
        "description" : "Returns all persons matching the optional search criteria.",
        "tags" : [ "Persons" ],
        "parameters" : [ {
          "name" : "birthMonth",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "birthYear",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "firstName",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "lastName",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Persons found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PersonDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create person",
        "description" : "Creates a new person.",
        "tags" : [ "Persons" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePersonRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Person created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PersonDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Person already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/persons/{id}" : {
      "put" : {
        "summary" : "Update person",
        "description" : "Updates an existing person.",
        "tags" : [ "Persons" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdatePersonRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Person updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PersonDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Person not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Person already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get" : {
        "summary" : "Find person by id",
        "description" : "Returns a single person by id.",
        "tags" : [ "Persons" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Person found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PersonDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Person not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete person",
        "description" : "Deletes a person and dependent person relations.",
        "tags" : [ "Persons" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Person deleted"
          },
          "404" : {
            "description" : "Person not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "info" : {
    "title" : "Coach Buddy API",
    "version" : "1.0.0",
    "description" : "API für Trainings-, Team- und Personenverwaltung"
  }
}